use prerelease version of teraslice-cli for prerelease asset workflow (#49)
Changes to teraslice-cli are needed for current dev assets to build
properly. This PR pins teraslice-cli to the prerelease tag, which is the
latest dev version on npm at any given time. We may have to remove this
in the future if making dev assets without a dev version of teraslice
being active. In that case we would be using some outdated teraslice-cli
version.